requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.4.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==3.0.5
pytz==2020.1
sqlparse==0.3.1
